+++ lib/dsql.dart
@@ -1,4 +1,12 @@
 library dsql;
 
+export 'src/filters/filter.dart';
+export 'src/filters/string_filter.dart';
+export 'src/filters/integer_filter.dart';
+export 'src/filters/double_filter.dart';
+export 'src/filters/boolean_filter.dart';
+export 'src/filters/timestamp_filter.dart';
+export 'src/dsql_utils.dart';
+
 export 'package:postgres/postgres.dart';
 export 'package:path/path.dart';