+++ lib/src/filters/filter.dart
@@ -1,0 +1,4 @@
+abstract class Filter<T extends Object> {
+  T get value;
+  String get operator;
+}